[Chore] #21 - 폴더링 위치 변경
